+++ apps/app/build.yaml
@@ -19,6 +19,7 @@ targets:
         generate_for:
           include:
             - lib/{provider,**/provider}/**.dart
+            - lib/router/router.dart
       # https://github.com/dart-lang/source_gen#ignore_for_file
       source_gen:combining_builder:
         options:
