+++ apps/app/lib/router/routes/main/main_page_shell_route.dart
@@ -1,4 +1,4 @@
-part of 'package:flutter_app/router/provider/router.dart';
+part of 'package:flutter_app/router/router.dart';
 
 @TypedStatefulShellRoute<MainPageShellRoute>(
   branches: [
