+++ apps/app/lib/router/routes/maintenance_page_route.dart
@@ -1,4 +1,4 @@
-part of 'package:flutter_app/router/provider/router.dart';
+part of 'package:flutter_app/router/router.dart';
 
 @TypedGoRoute<MaintenancePageRoute>(
   path: MaintenancePageRoute.path,
